Split organizationmemberteam and projectteam into their own files
